requirements.txt
pyttsx3==2.90
pillow==10.3.0